agentic_codebase_genius/BE/requirements.txt
byllm
jac-cloud
gitpython